+++ dart_native/ios/Classes/DNBlockWrapper.m
@@ -10,7 +10,6 @@
 #import "DNFFIHelper.h"
 #import "DNInvocation.h"
 #import <objc/runtime.h>
-#import "NSThread+DartNative.h"
 #import "DNPointerWrapper.h"
 #import "native_runtime.h"
 #import "DNError.h"
@@ -153,7 +152,7 @@ static atomic_uint_fast64_t _seq = 0;
         }
     }
     free(_typeEncodings);
-    NotifyDeallocToDart(_sequence, _dartPort);
+    NotifyDeallocToDart((intptr_t)_sequence, _dartPort);
 }
 
 - (void)initBlockWithError:(out NSError **)error {
@@ -382,9 +381,10 @@ static void DNFFIBlockClosureFunc(ffi_cif *cif, void *ret, void **args, void *us
     }
     retObjectAddr = (int64_t)*(void **)retAddr;
     DNHandleReturnValue(ret, wrapper, invocation);
-    [wrapper.thread dn_performBlock:^{
-        NSThread.currentThread.threadDictionary[@(retObjectAddr)] = nil;
-    }];
+    const char *type = wrapper.typeEncodings[0];
+    if (type == native_type_object || type == native_type_block) {
+        native_release_object((__bridge id)*(void **)retAddr);
+    }
 }