+++ dart_native/ios/Classes/DNMethodIMP.h
@@ -20,11 +20,9 @@ typedef void (*NativeMethodCallback)(void *_Nullable *_Null_unspecified args,
 
 @property (nonatomic, readonly) NativeMethodCallback callback;
 @property (nonatomic, getter=hasStret, readonly) BOOL stret;
-@property (nonatomic, readonly) Dart_Port dartPort;
 
 - (instancetype)initWithTypeEncoding:(const char *)typeEncodings
                             callback:(NativeMethodCallback)callback
-                            dartPort:(Dart_Port)dartPort
                                error:(NSError **)error;
 - (IMP)imp;