+++ dart_native/ios/Classes/DNObjectDealloc.m
@@ -6,70 +6,44 @@
 //
 
 #import "DNObjectDealloc.h"
-#import <objc/runtime.h>
 #import "native_runtime.h"
-
-#if !__has_feature(objc_arc)
-#error
-#endif
-
-#if TARGET_OS_OSX && __x86_64__
-    // 64-bit Mac - tag bit is LSB
-#   define OBJC_MSB_TAGGED_POINTERS 0
-#else
-    // Everything else - tag bit is MSB
-#   define OBJC_MSB_TAGGED_POINTERS 1
-#endif
-
-#if OBJC_MSB_TAGGED_POINTERS
-#   define _OBJC_TAG_MASK (1UL<<63)
-#else
-#   define _OBJC_TAG_MASK 1UL
-#endif
-
-static inline bool
-_objc_isTaggedPointer(const void *ptr) {
-    return ((uintptr_t)ptr & _OBJC_TAG_MASK) == _OBJC_TAG_MASK;
-}
 
 @interface DNObjectDealloc ()
 
-@property (nonatomic, readonly, weak) NSObject *host;
-@property (nonatomic, readonly) int64_t hostAddress;
-@property (nonatomic, readonly) Dart_Port dartPort;
+@property (nonatomic, readonly) intptr_t hostAddress;
 
 @end
 
 @implementation DNObjectDealloc
 
-+ (void)attachHost:(NSObject *)host
-          dartPort:(Dart_Port)dartPort {
-    if (!host || objc_getAssociatedObject(host, @selector(initWithHost:dartPort:))) {
-        return;
-    }
-    if (!_objc_isTaggedPointer((__bridge const void *)(host)) ||
-        [host isKindOfClass:NSClassFromString(@"__NSMallocBlock")]) {
-        __unused DNObjectDealloc *dealloc = [[self alloc] initWithHost:host
-                                                              dartPort:dartPort];
-    }
-}
-
-static int64_t fuxk;
+static const void *DNObjectDeallocStorageKey = (void *)&DNObjectDeallocStorageKey;
 
 - (instancetype)initWithHost:(NSObject *)host
-                    dartPort:(Dart_Port)dartPort {
-    self = [super init];
+                  storageKey:(nonnull const void *)storageKey {
+    self = [super initWithHost:host storageKey:storageKey];
     if (self) {
-        _host = host;
-        _hostAddress = (int64_t)host;
-        _dartPort = dartPort;
-        objc_setAssociatedObject(host, _cmd, self, OBJC_ASSOCIATION_RETAIN_NONATOMIC);
+        _hostAddress = (intptr_t)host;
     }
     return self;
 }
 
++ (instancetype)objectForHost:(NSObject *)host {
+    return [super objectForHost:host storageKey:DNObjectDeallocStorageKey];
+}
+
++ (nullable instancetype)attachHost:(NSObject *)host
+                           dartPort:(Dart_Port)dartPort {
+    DNObjectDealloc *dealloc = [super attachHost:host
+                                        dartPort:dartPort
+                                      storageKey:DNObjectDeallocStorageKey];
+    return dealloc;
+}
+
 - (void)dealloc {
-    NotifyDeallocToDart(_hostAddress, _dartPort);
+    NSSet<NSNumber *> *dartPorts = self.dartPorts;
+    for (NSNumber *dartPort in dartPorts) {
+        NotifyDeallocToDart(_hostAddress, dartPort.integerValue);
+    }
 }
 
 @end