+++ dart_native/lib/src/ios/common/callback_register.dart
@@ -84,7 +84,7 @@ _callback(
     }
   }
   if (result is id) {
-    markAutoreleasereturnObject(result.pointer);
+    retainObject(result.pointer);
   }
 }