+++ dart_native/lib/src/ios/runtime/block.dart
@@ -234,7 +234,7 @@ _callback(Pointer<Pointer<Pointer<Void>>> argsPtrPtrPtr,
     }
   }
   if (result is id) {
-    markAutoreleasereturnObject(result.pointer);
+    retainObject(result.pointer);
   }
 }