+++ dart_native/lib/src/ios/runtime/internal/native_runtime.dart
@@ -126,9 +126,9 @@ typedef BlockInvokeD = Pointer<Void> Function(
 final BlockInvokeD blockInvoke = runtimeLib
     .lookupFunction<BlockInvokeC, BlockInvokeD>('native_block_invoke');
 
-final void Function(Pointer<Void>) markAutoreleasereturnObject = nativeDylib
+final void Function(Pointer<Void>) retainObject = nativeDylib
     .lookup<NativeFunction<Void Function(Pointer<Void>)>>(
-        'native_mark_autoreleasereturn_object')
+        'native_retain_object')
     .asFunction();
 
 typedef ConvertNSStringToUTF16 = Pointer<Void> Function(
@@ -137,7 +137,7 @@ final ConvertNSStringToUTF16 convertNSStringToUTF16 =
     nativeDylib.lookupFunction<ConvertNSStringToUTF16, ConvertNSStringToUTF16>(
         'native_convert_nsstring_to_utf16');
 
-final void Function(Object, Pointer<Void>) passObjectToC = nativeDylib
-    .lookup<NativeFunction<Void Function(Handle, Pointer<Void>)>>(
+final int Function(Object, Pointer<Void>) passObjectToC = nativeDylib
+    .lookup<NativeFunction<Int32 Function(Handle, Pointer<Void>)>>(
         "PassObjectToCUseDynamicLinking")
     .asFunction();